refactor: move simple sampling method to sampling module
