+++ src/pysatl_core/distributions/strategies.py
@@ -13,10 +13,8 @@ __license__ = "SPDX-License-Identifier: MIT"
 
 from typing import TYPE_CHECKING, Any, Protocol, cast
 
-import numpy as np
-
 from pysatl_core.distributions.registry import characteristic_registry
-from pysatl_core.types import CharacteristicName, Method, NumericArray
+from pysatl_core.types import Method, NumericArray
 
 if TYPE_CHECKING:
     from collections.abc import Mapping
@@ -236,47 +234,3 @@ class SamplingStrategy(Protocol):
     """Protocol for strategies that generate samples from distributions."""
 
     def sample(self, n: int, distr: Distribution, **options: Any) -> NumericArray: ...
-
-
-class DefaultSamplingUnivariateStrategy(SamplingStrategy):
-    """
-    Default univariate sampler based on inverse transform sampling.
-
-    This strategy generates samples by applying the PPF (inverse CDF)
-    to uniformly distributed random variables.
-
-    Notes
-    -----
-    - Requires the distribution to provide a PPF computation method.
-    - Assumes that the PPF follows NumPy semantics (vectorized evaluation).
-    - Graph-derived PPFs (scalar-only) are currently not supported.
-    - Returns a NumPy array containing the generated samples.
-    """
-
-    def sample(self, n: int, distr: Distribution, **options: Any) -> NumericArray:
-        """
-        Generate samples from the distribution.
-
-        Parameters
-        ----------
-        n : int
-            Number of samples to generate.
-        distr : Distribution
-            Distribution to sample from.
-        **options : Any
-            Additional options forwarded to the PPF computation.
-
-        Returns
-        -------
-        NumericArray
-            NumPy array containing ``n`` generated samples.
-            The exact array shape depends on the distribution and sampling strategy.
-        """
-        ppf = distr.query_method(CharacteristicName.PPF, **options)
-        rng = np.random.default_rng()
-        U = rng.random(n)
-        # TODO: Now it will be based on the fact that the characteristic
-        #  has NumPy semantics (It is much more faster), that is,
-        #  it will not work with the graph computed characteristics currently.
-        samples = ppf(U)
-        return cast(NumericArray, samples)
