+++ src/pysatl_core/sampling/__init__.py
@@ -6,6 +6,7 @@ __author__ = "Artem Romanyuk"
 __copyright__ = "Copyright (c) 2025 PySATL project"
 __license__ = "SPDX-License-Identifier: MIT"
 
+from .default import DefaultSamplingUnivariateStrategy
 from .unuran import (
     DefaultUnuranSampler,
     DefaultUnuranSamplingStrategy,
@@ -19,6 +20,7 @@ DefaultSampler = DefaultUnuranSampler
 DefaultSamplingStrategy = DefaultUnuranSamplingStrategy
 
 __all__ = [
+    "DefaultSamplingUnivariateStrategy",
     "SamplingMethod",
     "SamplingMethodConfig",
     "DefaultSampler",
